+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Added some additional compatibility logic to make styles with old corridor definitions work in some additional cases.
 * Switch more package loading errors to be displayed in the more friendly error window.
 * Add support for Connection Signage producing blue/orange frames.
+* Added ability to use `<marker>` for filenames in editoritems, to generate a unique file 
+  for cases where instances should be replaced by conditions.
 
 # Version 4.46.1
 * Fix FGD export producing invalid syntax.